src/perf_tests/requirements.txt
locust==2.20.1
prometheus-client==0.19.0
python-ulid==2.2.0
